Move logging settings to proper module
